+++ packages/canvas_editor_flutter/test/image_import_extension_test.dart
@@ -88,9 +88,7 @@ final class _DeferredImageImportPort implements ImageImportPort {
       Completer<ImageImportResult>();
 
   @override
-  Future<ImageImportResult> importImage({
-    required ImageImportSource source,
-  }) {
+  Future<ImageImportResult> importImage({required ImageImportSource source}) {
     requestedSources.add(source);
     return _completer.future;
   }